Update popup.html
upon reviewing documentation after this broke, changed resolution as it was below the minimum required resolution.  That didn't fix it, but for some reason changing the domain to from example to twitter lets it run fine across any tab, including just the new tab.  Dunno why that fixed it, but it works now??